// src/typings/api.ts
export interface ApiSubscriptionProductFamily {
    name: string
    subscriptionProducts: ApiSubscriptionProduct[]
}

export interface ApiSubscriptionProduct {
    subscriptionProductId: number
    name: string
    dependencies: ApiSubscriptionProductDependency
    subscriptionType: string
}

export interface ApiSubscriptionProductDependency {
    automaticallyAdded: ApiAddOnProductLink[]
    options: ApiSubscriptionProductDependencyOption[]
}

export interface ApiSubscriptionProductDependencyOption {
    numberOfProductsToSelect: number,
    products: ApiAddOnProductLink[]
}

export interface ApiAddOnProductLink {
    productId: number
    productName: string
}